# Faster and Better weak generators