I took a couple of days to research some fast pseudo-random number generators and I thought it'd be nice to contribute to the crate.

This PR contains two commits one adding the xorshift128+ and the other PCG RXS M XS (32 and 64bit variants). I don't think there's precedence for adding both, so I think we should discuss how to proceed.

| Algorithm    | Size     | Quality | x86 (u32) | x86 (u64) | x64 (u32) | x64 (u64) |
|--------------|----------|---------|-----------|-----------|-----------|-----------|
| PCG32        | 4 Bytes  | Fair    | 1365      | 1616      | 1646      | 1677      |
| PCG64        | 8 Bytes  | Great   | 510       | 1016      | 1646      | 3493      |
| Xorshift128   | 16 Bytes | Fair     | 1486      | 1895      | 2094      | 2279      |
| Xorshift128+ | 16 Bytes | Great   | 911       | 1656      | 2000      | 4278      |

x64 benchmarks (my laptop)
```
test algorithms::chacha_u32                        ... bench:       2,477 ns/iter (+/- 191) = 161 MB/s
test algorithms::chacha_u64                        ... bench:       4,783 ns/iter (+/- 353) = 167 MB/s
test algorithms::isaac64_u32                       ... bench:         617 ns/iter (+/- 40) = 648 MB/s
test algorithms::isaac64_u64                       ... bench:         639 ns/iter (+/- 16) = 1251 MB/s
test algorithms::isaac_u32                         ... bench:         847 ns/iter (+/- 30) = 472 MB/s
test algorithms::isaac_u64                         ... bench:       1,579 ns/iter (+/- 130) = 506 MB/s
test algorithms::pcg32_u32                         ... bench:         243 ns/iter (+/- 10) = 1646 MB/s
test algorithms::pcg32_u64                         ... bench:         477 ns/iter (+/- 23) = 1677 MB/s
test algorithms::pcg64_u32                         ... bench:         243 ns/iter (+/- 14) = 1646 MB/s
test algorithms::pcg64_u64                         ... bench:         229 ns/iter (+/- 16) = 3493 MB/s
test algorithms::xorshift_u32                      ... bench:         191 ns/iter (+/- 11) = 2094 MB/s
test algorithms::xorshift_u64                      ... bench:         351 ns/iter (+/- 18) = 2279 MB/s
test algorithms::xorshiftp_u32                     ... bench:         200 ns/iter (+/- 11) = 2000 MB/s
test algorithms::xorshiftp_u64                     ... bench:         187 ns/iter (+/- 7) = 4278 MB/s

```
x86 benchmarks (DO)
```
test algorithms::chacha_u32                        ... bench:       2,644 ns/iter (+/- 21) = 151 MB/s
test algorithms::chacha_u64                        ... bench:       4,964 ns/iter (+/- 28) = 161 MB/s
test algorithms::isaac64_u32                       ... bench:       1,038 ns/iter (+/- 8) = 385 MB/s
test algorithms::isaac64_u64                       ... bench:       1,059 ns/iter (+/- 6) = 755 MB/s
test algorithms::isaac_u32                         ... bench:       1,107 ns/iter (+/- 12) = 361 MB/s
test algorithms::isaac_u64                         ... bench:       2,004 ns/iter (+/- 13) = 399 MB/s
test algorithms::pcg32_u32                         ... bench:         293 ns/iter (+/- 4) = 1365 MB/s
test algorithms::pcg32_u64                         ... bench:         495 ns/iter (+/- 6) = 1616 MB/s
test algorithms::pcg64_u32                         ... bench:         783 ns/iter (+/- 15) = 510 MB/s
test algorithms::pcg64_u64                         ... bench:         787 ns/iter (+/- 9) = 1016 MB/s
test algorithms::xorshift_u32                      ... bench:         269 ns/iter (+/- 4) = 1486 MB/s
test algorithms::xorshift_u64                      ... bench:         422 ns/iter (+/- 6) = 1895 MB/s
test algorithms::xorshiftp_u32                     ... bench:         439 ns/iter (+/- 7) = 911 MB/s
test algorithms::xorshiftp_u64                     ... bench:         483 ns/iter (+/- 5) = 1656 MB/s
```

PCG is neat, it's very compact and provide good quality, even in the 32bit variant!

Although my preference would be to add the xorshift128+ variant and leave the xorshift128 there (probably worth a rename though).

In addition weak_rng should return the fastest algorithm available for the architectures. As 64bit operations kills performance in 32bit architectures this can be as simple as Xorshift64 for 32bit and Xorshift128+ for 64bit.

Finally, users won't need to look elsewhere for a crazy fast random implementation.

Thoughts?