@nagisa 2015-07-17
Hi!

Xorshift+ sounds like a sensible addition. As far as PCG is concerned, I’d implement it in an external crate (e.g. .

> In addition weak_rng should return the fastest algorithm available for the platform.

The “for the platform” is a pretty hard thing to do ­– we don’t want to start benchmarking all the xorshifts at runtime, nor do we want to make weak_rng to return XorShift on some platforms and XorShiftP on others. Therefore we have to rely on eyeballing and pick an algorithm that seems to be fast(est) on as many platforms as possible (our current pick being XorShift). 

Changing the rng returned by `weak_rng`, sadly, is also a somewhat breaking change (breaks code such as `let rng: XorShift = weak_rng()`) and a small increase of performance on some of the workloads does not sound convincing enough to make the change. What *does* sound, convincing, though, is that XorShift+ passes BigCrush (as per Wikipedia), while our current choice does not.