@arthurprs 2015-07-17
@nagisa by "fastest algorithm available for the platform." I meant, that 64bits platform should default to Xorshift128+ as it's better overall, both quality and speed wise.

I agree that it'll break some code (not much though) but I'd argue that it's both worthy and expected as per semver.