@alexcrichton 2015-07-17
r? @huonw, you may have more opinions on this than I

> The “for the platform” is a pretty hard thing to do ­– we don’t want to start benchmarking all the xorshifts at runtime, nor do we want to make weak_rng to return XorShift on some platforms and XorShift+ on others. Therefore we have to rely on eyeballing and pick an algorithm that seems to be fast(est) on as many platforms as possible (our current pick being XorShift). 

While I agree we don't want to determine the fastest one at runtime, I do think that we should change the return type of `weak_rng` to `WeakRng` to hide the implementation. That way we can swap it out on various platforms and also change it in the future (like we want to right now) without breaking code. 

For now I'd avoid changing `weak_rng` as this crate will likely have a number of other breaking changes coming soon, and it probably isn't worth the breakage just yet.