@huonw 2015-07-17
> leave the xorshift64 there

Erm, maybe we're using different notation, but I thought the generator implemented here was xorshift128.