@arthurprs 2015-07-17
Updated the PR code to
* avoid breaking changes for weak_rng (will submit separately)
* tests for panicking for 0 seed
* renamed structs
* added more doc lines for structs
* what we have right now is xorshift128 not xorshift64... sorry everyone

@huonw yaaa, somehow 64 sticked into my head. Thank you.