@DanielKeep 2015-07-18
@nagisa There's [one by EricIO](https://github.com/EricIO/pcg-rust) and [one by codahale](https://github.com/codahale/pcg), which look to be minimal ports of Pcg.  I just started a [full port of the entire C++ reference implementation](https://github.com/DanielKeep/rust-pcg), mostly as an exercise in porting over C++ code using template mixins.

I aim to also port the crazy "extended" generators that allow for generators with arbitrarily large periods... y'know, *because it's there*.

Any input in relation to *this* crate is welcome.