@arthurprs 2015-07-23
Just to make it clearer, I removed all breaking changes on this.
So what's in is the new xorshift128+ implementation, tests for xorshift variants and code/benchmarks cleanup.

We can integrate the Weak abstraction (to hide the implementation) in a subsequent pr/version.