@arthurprs 2015-08-04
Hello again, any updates on this? I believe this is a good addition as it is.

It'd be specially useful if we hid the implementation of weak_rng() in a WeakRng struct, this way we could hide the fastest implementation behind it. But this is a breaking change and we must make a decision.

Anyway I believe this should be merged to encourage further improvements down the road (like the previous mentioned WeakRng struct).