@zackw 2015-09-15
I would like to stick my oar in: Please do NOT provide any "weak" algorithms.  Once they are in, they can't be got rid of because RNGs need to be reproducible even decades later; but bad randomness is an endless well of subtle bugs, often security-critical bugs.

I would hesitate even to provide ISAAC.