@zackw 2015-09-15
@pcwalton Why do you want to discourage the use of userspace CSPRNGs?  I ask largely because people have been making noises about adding ChaCha-based `arc4random*` to glibc, and I'm not aware of any reason why that would be a bad idea.  (No end of implementation headches, yes, but.)