@bhickey 2015-10-24
After chatting with some other RNG enthusiasts I'd advocate for removing non-crypto PRNGs from `rand`. Ideally I'd fork `rand` into `rand-crypt` and `rand-sim`, but I think this is unlikely to gain traction. There are very different design requirements for simulation and crypto and we may be ill served by cramming them together. For games and simulation reproducibility is very important, while security demands easily verified code. Sequestering simulation-quality PRNGs to another crate simplifies verification (ex. If you're using `rand-sim`, you're provably insecure) and reduces conflict between competing goals. @bstrie @robojeb