@DanielKeep 2015-10-25
Couldn't this split also be achieved by having two top-level modules within the crate: `secure` and `weak`?  You would probably want to remove top-level aliases (other than a "I don't care, just gimme random values" that defaults to something secure), which means *any* attempt to use a generator involves explicitly choosing between secure and weak generators.