@bhickey 2015-10-25
@arthurprs There's no reason the interface would need to change. If we forked the crate today Rng would go on being the common trait.
 
@DanielKeep `rand` doesn't make any promises about reproducibility. Changes to support code can and will break anyone who wants to do repeatable simulation.